backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0  # Async PostgreSQL support

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Integration (Self-hosted friendly)
httpx==0.26.0
aiofiles==23.2.1

# Free AI Options (Self-hosted)
# ollama>=0.1.0  # Uncomment for local AI via Ollama
# transformers>=4.36.0  # Uncomment for HuggingFace local models

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Generation (Receipts) - Free/Open Source
reportlab==4.0.8

# Async
anyio==4.2.0
asyncio==3.4.3

# SMS Services - Paid (with free test mode)
twilio>=9.0.0

# Email-to-SMS Support - Free
aiosmtplib>=3.0.0
